퍼센테이지에 소수점으로 나와서 급수정

+++ app/src/main/java/com/usungsoft/usLibrary/views/UsProgressView.java
@@ -182,7 +182,7 @@ public class UsProgressView extends ConstraintLayout {
                             try {
                                 int percentage = valueSum / mReadedValues.size();
                                 mProgressView.setProgress(percentage * mUpdatingFloatValue);
-                                mTvPercentage.setText((percentage * (mUpdatingFloatValue * 100)) + "%");
+                                mTvPercentage.setText((int) (percentage * (mUpdatingFloatValue * 100)) + "%");
                                 mReadedValues.clear();
                             } catch (ArithmeticException e) {  }
                         });
